refactor: rename browserslist-config-autoguru to browserslist-config and update related paths
